refactor: pull rankings generation out into a separate file
affects: tymly-rankings-plugin
