+++ plugins/tymly-rankings-plugin/lib/components/services/rankings/to-two-dp.js
@@ -1,0 +1,5 @@
+function toTwoDp(num) {
+  return Math.round(num*100)/100
+} // toTwoDp
+
+module.exports = toTwoDp
